+++ NEWS.md
@@ -1,4 +1,19 @@
-## [zen4R 0.7-1](https://cran.r-project.org/package=zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-red.svg)](https://cran.r-project.org/package=zen4R)
+# Change log
+
+## [zen4R 0.8](https://cran.r-project.org/package=zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=zen4R)
+
+**Enhancements**
+
+* [#106](https://github.com/eblondel/zen4R/issues/106) Quote DOIs in deposit/record search methods
+* [#108](https://github.com/eblondel/zen4R/issues/108) `getFunders` additional `q` Elastic search argument
+* [#109](https://github.com/eblondel/zen4R/issues/109) `getGrants` additional `q` Elastic search argument
+
+**New features**
+
+* [#110](https://github.com/eblondel/zen4R/issues/110) `getFundersByName` new method
+* [#111](https://github.com/eblondel/zen4R/issues/111) `getGrantsByName` new method
+
+## [zen4R 0.7-1](https://cran.r-project.org/package=zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=zen4R)
 
 **Bug fixes**
 
@@ -6,7 +21,7 @@
 * [#102](https://github.com/eblondel/zen4R/issues/102) User-Agent related issue with `download_zenodo` - relates to #103
 * [#103](https://github.com/eblondel/zen4R/issues/103) Change User-Agent to bypass Zenodo API server restrictions
 
-## [zen4R 0.7](https://cran.r-project.org/package=zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-red.svg)](https://cran.r-project.org/package=zen4R)
+## [zen4R 0.7](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.7.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.7.tar.gz)
 
 **Bug fixes**
 
@@ -27,7 +42,7 @@
 
 * [#97](https://github.com/eblondel/zen4R/issues/97) Upgrade roxygen2 7.2.1 to fix html issues
 
-## [zen4R 0.6-1](https://cran.r-project.org/package=zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=zen4R)
+## [zen4R 0.6-1](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.6-1.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.6-1.tar.gz)
 
 **Bug fixes**